refactor: move WKN to its file, update imports
