+++ crates/miden-lib/Cargo.toml
@@ -28,7 +28,6 @@ thiserror = { workspace = true }
 [dev-dependencies]
 miden-objects = { workspace = true, features = ["testing"] }
 vm-processor = { workspace = true, features = ["testing"] }
-assembly = { workspace = true, features = ["testing"] }
 
 [build-dependencies]
 assembly = { workspace = true }
