+++ crates/miden-lib/src/note/scripts.rs
@@ -33,7 +33,7 @@ pub fn p2id() -> NoteScript {
 
 /// Returns the P2ID (Pay-to-ID) note script commitment.
 pub fn p2id_commitment() -> Digest {
-    P2ID_SCRIPT.clone().hash()
+    P2ID_SCRIPT.hash()
 }
 
 /// Returns the P2IDR (Pay-to-ID with recall) note script.
@@ -43,7 +43,7 @@ pub fn p2idr() -> NoteScript {
 
 /// Returns the P2IDR (Pay-to-ID with recall) note script commitment.
 pub fn p2idr_commitment() -> Digest {
-    P2IDR_SCRIPT.clone().hash()
+    P2IDR_SCRIPT.hash()
 }
 
 /// Returns the SWAP (Swap note) note script.
@@ -53,5 +53,5 @@ pub fn swap() -> NoteScript {
 
 /// Returns the SWAP (Swap note) note script commitment.
 pub fn swap_commitment() -> Digest {
-    SWAP_SCRIPT.clone().hash()
+    SWAP_SCRIPT.hash()
 }
